extract ticket sales config to new files
